requirements.txt
# Core dependencies
ollama>=0.1.0
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0

# Audio processing (optional for basic functionality)
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudio>=0.2.11

# ML/AI utilities (optional for enhanced features)
openai-whisper>=20231117
torch>=2.0.0
transformers>=4.30.0

# TTS (optional - may require specific setup)
# TTS>=0.22.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0